+++ pkg/state_handler/play_controller.go
@@ -11,5 +11,4 @@ type Controller interface {
 	CtrlPrevious()
 	CtrlSeek(duration time.Duration)
 	CtrlSetVolume(volume int)
-	PlayingInfo() PlayingInfo
 }
