+++ pkg/ui/player_controller.go
@@ -2,64 +2,52 @@ package ui
 
 import (
 	"time"
-
-	"github.com/go-musicfox/go-musicfox/pkg/state_handler"
 )
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlPaused() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlPaused}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlResume() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlResume}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlStop() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlStop}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlToggle() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlToggle}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlNext() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlNext}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlPrevious() {
 	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.ctrl <- CtrlSignal{Type: CtrlPrevious}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlSeek(duration time.Duration) {
-	p.ctrl <- CtrlSignal{
-		Type:     CtrlSeek,
-		Duration: duration,
-	}
+	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
+	p.ctrl <- CtrlSignal{Type: CtrlSeek, Duration: duration}
 }
 
+// Deprecated: Only state_handler.Handler can call this method, others please use Player instead.
 func (p *Player) CtrlSetVolume(volume int) {
-	// 不更新playingInfo
+	// NOTICE: 提供给state_handler调用，因为有GC panic问题，这里使用chan传递
 	p.Player.SetVolume(volume)
-}
-
-func (p *Player) PlayingInfo() state_handler.PlayingInfo {
-	music := p.curSong
-	return state_handler.PlayingInfo{
-		TotalDuration:  music.Duration,
-		PassedDuration: p.PassedTime(),
-		State:          p.State(),
-		Volume:         p.Volume(),
-		TrackID:        music.Id,
-		PicUrl:         music.PicUrl,
-		Name:           music.Name,
-		Album:          music.Album.Name,
-		Artist:         music.ArtistName(),
-		AlbumArtist:    music.Album.ArtistName(),
-	}
 }
